test: split storage tests
