+++ test/storage.test.js
@@ -274,97 +274,6 @@ test("Sending messages with a name Storage should use the name as source", t => 
   ]);
 });
 
-test("Sending messages on a Storage should send them to matching subscribers", t => {
-  const s = new Storage();
-  const emit = t.context.spy(s, "emit");
-  const recv1 = t.context.stub();
-  const recv2 = t.context.stub();
-  const msg = { tag: "testMessage" };
-  const msg2 = { tag: "uncaught" };
-
-  s.addEffect({ effect: recv1, subscribe: { "testMessage": true } });
-  s.addEffect({ effect: recv2, subscribe: { "fooMessage": true } });
-
-  s.sendMessage(msg);
-
-  t.deepEqual(args(recv1), [[msg, ["$"]]]);
-  t.deepEqual(args(recv2), []);
-  t.deepEqual(args(emit), [
-    ["messageQueued", msg, ["$"]],
-    ["messageMatched", msg, [], false],
-  ]);
-
-  s.sendMessage(msg2);
-
-  t.deepEqual(args(recv1), [[msg, ["$"]]]);
-  t.deepEqual(args(recv2), []);
-  t.deepEqual(args(emit), [
-    ["messageQueued", msg, ["$"]],
-    ["messageMatched", msg, [], false],
-    ["messageQueued", msg2, ["$"]],
-    ["unhandledMessage", msg2, ["$"]],
-  ]);
-});
-
-test("Sending messages with a name on a Storage should propagate the name", t => {
-  const s = new Storage();
-  const emit = t.context.spy(s, "emit");
-  const recv1 = t.context.stub();
-  const msg = { tag: "testMessage" };
-
-  s.addEffect({ effect: recv1, subscribe: { testMessage: true } });
-
-  s.sendMessage(msg, "themessage");
-
-  t.deepEqual(args(recv1), [[msg, ["themessage"]]]);
-  t.deepEqual(args(emit), [
-    ["messageQueued", msg, ["themessage"]],
-    ["messageMatched", msg, [], false],
-  ]);
-});
-
-test("Removing a subscriber should not fire when a matching message is sent", t => {
-  const s = new Storage();
-  const emit = t.context.spy(s, "emit");
-  const recv1 = t.context.stub();
-  const recv2 = t.context.stub();
-  const msg = { tag: "testMessage" };
-
-  const eff1 = { effect: recv1, subscribe: { testMessage: true } };
-  const eff2 = { effect: recv2, subscribe: { testMessage: true } };
-
-  s.addEffect(eff1);
-  s.addEffect(eff2);
-  s.removeEffect(eff2);
-
-  s.sendMessage(msg);
-
-  t.deepEqual(args(emit), [
-    ["messageQueued", msg, ["$"]],
-    ["messageMatched", msg, [], false],
-  ]);
-  t.deepEqual(args(recv1), [[msg, ["$"]]]);
-  t.deepEqual(args(recv2), []);
-});
-
-test("Sending messages on a Storage should also trigger unhandledMessage if no active subscribers are present", t => {
-  const s = new Storage();
-  const emit = t.context.spy(s, "emit");
-  const recv = t.context.stub();
-  const msg = { tag: "testMessage" };
-
-  s.addEffect({ effect: recv, subscribe: { "testMessage": { passive: true } } });
-
-  s.sendMessage(msg);
-
-  t.deepEqual(args(recv), [[msg, ["$"]]]);
-  t.deepEqual(args(emit), [
-    ["messageQueued", msg, ["$"]],
-    ["messageMatched", msg, [], true],
-    ["unhandledMessage", msg, ["$"]],
-  ]);
-});
-
 test("States with init using updateAndSend should send messages to parent Storage", t => {
   const s = new Storage();
   const emit = t.context.spy(s, "emit");
@@ -685,81 +594,6 @@ test("Messages generated during processing are handled in order", t => {
   ]);
 });
 
-test("Active subscribers prevent parents and unhandledMessage from receiving", t => {
-  const s = new Storage();
-  const emit = t.context.spy(s, "emit");
-  const stub1 = t.context.stub();
-  const stub2 = t.context.stub();
-  const firstData = { name: "firstData" };
-  const initMsg = { tag: "initMsg" };
-  const firstMsg = { tag: "firstMsg" };
-  const firstDef = {
-    id: "first",
-    init: t.context.stub(() => updateData(firstData)),
-    update: t.context.stub(() => updateAndSend(firstData, firstMsg)),
-    subscribe: t.context.stub(() => ({ "initMsg": true })),
-  };
-
-  s.addEffect({ effect: stub1, subscribe: { "initMsg": true } });
-  s.addEffect({ effect: stub2, subscribe: { "firstMsg": true } });
-
-  const first = s.createState(firstDef);
-
-  first.sendMessage(initMsg);
-
-  t.deepEqual(args(stub1), []);
-  t.deepEqual(args(stub2), [[firstMsg, ["first"]]]);
-  t.deepEqual(args(firstDef.update), [[firstData, initMsg]]);
-  t.is(first.getData(), firstData);
-  t.deepEqual(args(emit), [
-    ["stateCreated", ["first"], undefined, { name: "firstData" }],
-    ["messageQueued", { tag: "initMsg" }, ["first", "$"]],
-    ["messageMatched", { tag: "initMsg" }, ["first"], false],
-    ["stateNewData", { name: "firstData" }, ["first"], { tag: "initMsg" }],
-    ["messageQueued", { tag: "firstMsg" }, ["first"]],
-    ["messageMatched", { tag: "firstMsg" }, [], false],
-  ]);
-});
-
-test("Passive subscribers always receive messages from children", t => {
-  const s = new Storage();
-  const emit = t.context.spy(s, "emit");
-  const stub1 = t.context.stub();
-  const stub2 = t.context.stub();
-  const firstData = { name: "firstData" };
-  const initMsg = { tag: "initMsg" };
-  const firstMsg = { tag: "firstMsg" };
-  const firstDef = {
-    id: "first",
-    init: t.context.stub(() => updateData(firstData)),
-    update: t.context.stub(() => updateAndSend(firstData, firstMsg)),
-    subscribe: t.context.stub(() => ({ "initMsg": true })),
-  };
-
-  s.addEffect({ effect: stub1, subscribe: { "initMsg": { passive: true } } });
-  s.addEffect({ effect: stub2, subscribe: { "firstMsg": { passive: true } } });
-
-  const first = s.createState(firstDef);
-
-  first.sendMessage(initMsg);
-
-  t.deepEqual(args(stub1), [[initMsg, ["first", "$"]]]);
-  t.deepEqual(args(stub2), [[firstMsg, ["first"]]]);
-  t.deepEqual(args(firstDef.update), [[firstData, initMsg]]);
-  t.is(first.getData(), firstData);
-
-  t.deepEqual(args(emit), [
-    ["stateCreated", ["first"], undefined, { name: "firstData" }],
-    ["messageQueued", { tag: "initMsg" }, ["first", "$"]],
-    ["messageMatched", { tag: "initMsg" }, ["first"], false],
-    ["stateNewData", { name: "firstData" }, ["first"], { tag: "initMsg" }],
-    ["messageQueued", { tag: "firstMsg" }, ["first"]],
-    ["messageMatched", { tag: "initMsg" }, [], true],
-    ["messageMatched", { tag: "firstMsg" }, [], true],
-    ["unhandledMessage", { tag: "firstMsg" }, ["first"]],
-  ]);
-});
-
 test("findClosestSupervisor", t => {
   const defA = {
     id: "a",
@@ -968,133 +802,6 @@ test("Storage updates subscribe during processing when state data is updated", t
     ["unhandledMessage", { tag: "b", __no: true }, ["a", "test"]],
     ["unhandledMessage", { tag: "a", __no: true }, ["a", "test"]],
   ]);
-});
-
-test("restoreSnapshot throws if it cannot find a matching state definition", t => {
-  const s = new Storage();
-  const emit = t.context.spy(s, "emit");
-
-  t.throws(() => s.restoreSnapshot({ foo: { id: "bar", data: null, nested: {} } }), { message: "Missing model for state 'bar'." });
-
-  t.deepEqual(args(emit), [
-    ["snapshotRestore", { foo: { id: "bar", data: null, nested: {} } }],
-  ]);
-});
-
-test("restoreSnapshot works on empty", t => {
-  const s = new Storage();
-  const emit = t.context.spy(s, "emit");
-
-  t.deepEqual(s.getSnapshot(), { });
-
-  s.restoreSnapshot({ });
-
-  t.deepEqual(s.getSnapshot(), { });
-
-  t.deepEqual(args(emit), [
-    ["snapshotRestore", {}],
-    ["snapshotRestored"],
-  ]);
-});
-
-test("restoreSnapshot restores a snapshot", t => {
-  const s = new Storage();
-  const emit = t.context.spy(s, "emit");
-  const init = t.context.stub(() => updateData(1));
-  const update = t.context.stub(() => updateData(2));
-  const subscribe = t.context.stub(() => ({}));
-  const d = {
-    id: "foo",
-    init,
-    update,
-    subscribe,
-  };
-
-  s.addModel(d);
-
-  t.deepEqual(s.getSnapshot(), { });
-
-  s.restoreSnapshot({ foo: { id: "foo", data: 3, nested: {} } });
-
-  t.deepEqual(s.getSnapshot(), { foo: { id: "foo", data: 3, nested: {} } });
-
-  t.not(s.getState(d), undefined);
-  t.deepEqual((s.getState(d): any).getData(), 3);
-
-  t.deepEqual(args(emit), [
-    ["snapshotRestore", { foo: { id: "foo", data: 3, nested: {} } }],
-    ["snapshotRestored"],
-  ]);
-  t.deepEqual(args(init), []);
-  t.deepEqual(args(update), []);
-  t.deepEqual(args(subscribe), []);
-});
-
-test("restoreSnapshot restores a snapshot with a differing name", t => {
-  const s = new Storage();
-  const emit = t.context.spy(s, "emit");
-  const init = t.context.stub(() => updateData(1));
-  const update = t.context.stub(() => updateData(2));
-  const subscribe = t.context.stub(() => ({}));
-  const d = {
-    id: "bar",
-    init,
-    update,
-    subscribe,
-  };
-
-  s.addModel(d);
-
-  t.deepEqual(s.getSnapshot(), { });
-
-  s.restoreSnapshot({ foo: { id: "bar", data: 3, nested: {} } });
-
-  t.deepEqual(s.getSnapshot(), { foo: { id: "bar", data: 3, nested: {} } });
-
-  t.not(s.getState(d, "foo"), undefined);
-
-  t.deepEqual(args(emit), [
-    ["snapshotRestore", { foo: { id: "bar", data: 3, nested: {} } }],
-    ["snapshotRestored"],
-  ]);
-  t.deepEqual(args(init), []);
-  t.deepEqual(args(update), []);
-  t.deepEqual(args(subscribe), []);
-});
-
-test("restoreSnapshot restores nested snapshots", t => {
-  const s = new Storage();
-  const emit = t.context.spy(s, "emit");
-  const init = t.context.stub(() => updateData(1));
-  const update = t.context.stub(() => updateData(2));
-  const subscribe = t.context.stub(() => ({}));
-  const d = {
-    id: "foo",
-    init,
-    update,
-    subscribe,
-  };
-
-  s.addModel(d);
-
-  t.deepEqual(s.getSnapshot(), { });
-
-  s.restoreSnapshot({ foo: { id: "foo", data: 3, nested: { foo: { id: "foo", data: 5, nested: {} } } } });
-
-  t.deepEqual(s.getSnapshot(), { foo: { id: "foo", data: 3, nested: { foo: { id: "foo", data: 5, nested: {} } } } });
-
-  t.not(s.getState(d), undefined);
-  t.deepEqual((s.getState(d): any).getData(), 3);
-  t.not((s.getState(d): any).getState(d), undefined);
-  t.deepEqual(((s.getState(d): any).getState(d): any).getData(), 5);
-
-  t.deepEqual(args(emit), [
-    ["snapshotRestore", { foo: { id: "foo", data: 3, nested: { foo: { id: "foo", data: 5, nested: {} } } } }],
-    ["snapshotRestored"],
-  ]);
-  t.deepEqual(args(init), []);
-  t.deepEqual(args(update), []);
-  t.deepEqual(args(subscribe), []);
 });
 
 test("Storage getState, createState, and removeState, with a different name should not affect the existing", t => {
@@ -1344,267 +1051,4 @@ test("createState with a different name should still work to send messages", t =
   t.is(init.calls.length, 3);
   t.deepEqual(args(update), []);
   t.is(subscribe.calls.length, 4);
-});
-
-test("broadcastMessage triggers unhandledMessage on empty", t => {
-  const s = new Storage();
-  const emit = t.context.spy(s, "emit");
-
-  s.broadcastMessage({ tag: "AAA" });
-
-  t.deepEqual(args(emit), [
-    ["messageQueued", { tag: "AAA" }, ["@"]],
-    ["unhandledMessage", { tag: "AAA" }, ["@"]],
-  ]);
-});
-
-test("broadcastMessage triggers attempts to send messages to all states", t => {
-  const s = new Storage();
-  const emit = t.context.spy(s, "emit");
-  const init = t.context.stub(() => updateData(1));
-  const update = t.context.stub(() => updateData(2));
-  const subscribe = t.context.stub(() => ({}));
-  const d = {
-    id: "foo",
-    init,
-    update,
-    subscribe,
-  };
-
-  const foo = s.createState(d);
-  foo.createState(d, undefined, "bar");
-
-  s.broadcastMessage({ tag: "AAA" });
-
-  t.deepEqual(args(emit), [
-    ["stateCreated", ["foo"], undefined, 1],
-    ["stateCreated", ["foo", "bar"], undefined, 1],
-    ["messageQueued", { tag: "AAA" }, ["@"]],
-    ["unhandledMessage", { tag: "AAA" }, ["@"]],
-  ]);
-  t.deepEqual(args(subscribe), [
-    [1],
-    [1],
-  ]);
-  t.deepEqual(args(update), []);
-});
-
-test("broadcastMessage sends a message to all states with deepest first", t => {
-  const s = new Storage();
-  const emit = t.context.spy(s, "emit");
-  const init = t.context.stub(() => updateData(1));
-  const update = t.context.stub(() => updateData(2));
-  const subscribe = t.context.stub(() => ({ AAA: true }));
-  const d = {
-    id: "foo",
-    init,
-    update,
-    subscribe,
-  };
-
-  const foo = s.createState(d);
-  foo.createState(d);
-  foo.createState(d, undefined, "bar");
-
-  s.broadcastMessage({ tag: "AAA" });
-
-  t.deepEqual(args(emit), [
-    ["stateCreated", ["foo"], undefined, 1],
-    ["stateCreated", ["foo", "foo"], undefined, 1],
-    ["stateCreated", ["foo", "bar"], undefined, 1],
-    ["messageQueued", { tag: "AAA" }, ["@"]],
-    ["messageMatched", { tag: "AAA" }, ["foo", "foo"], false],
-    ["stateNewData", 2, ["foo", "foo"], { tag: "AAA" }],
-    ["messageMatched", { tag: "AAA" }, ["foo", "bar"], false],
-    ["stateNewData", 2, ["foo", "bar"], { tag: "AAA" }],
-    ["messageMatched", { tag: "AAA" }, ["foo"], false],
-    ["stateNewData", 2, ["foo"], { tag: "AAA" }],
-  ]);
-  t.deepEqual(args(subscribe), [
-    [1],
-    [2],
-    [1],
-    [2],
-    [1],
-    [2],
-  ]);
-  t.is(update.calls.length, 3);
-});
-
-test("broadcastMessage will still trigger unhandledMessage if only passive subscribers are used", t => {
-  const s = new Storage();
-  const emit = t.context.spy(s, "emit");
-  const init = t.context.stub(() => updateData(1));
-  const update = t.context.stub(() => updateData(2));
-  const subscribe = t.context.stub(() => ({ AAA: { passive: true } }));
-  const d = {
-    id: "foo",
-    init,
-    update,
-    subscribe,
-  };
-
-  const foo = s.createState(d);
-  foo.createState(d);
-  foo.createState(d, undefined, "bar");
-
-  s.broadcastMessage({ tag: "AAA" });
-
-  t.deepEqual(args(emit), [
-    ["stateCreated", ["foo"], undefined, 1],
-    ["stateCreated", ["foo", "foo"], undefined, 1],
-    ["stateCreated", ["foo", "bar"], undefined, 1],
-    ["messageQueued", { tag: "AAA" }, ["@"]],
-    ["messageMatched", { tag: "AAA" }, ["foo", "foo"], true],
-    ["stateNewData", 2, ["foo", "foo"], { tag: "AAA" }],
-    ["messageMatched", { tag: "AAA" }, ["foo", "bar"], true],
-    ["stateNewData", 2, ["foo", "bar"], { tag: "AAA" }],
-    ["messageMatched", { tag: "AAA" }, ["foo"], true],
-    ["stateNewData", 2, ["foo"], { tag: "AAA" }],
-    ["unhandledMessage", { tag: "AAA" }, ["@"]],
-  ]);
-  t.deepEqual(args(subscribe), [
-    [1],
-    [2],
-    [1],
-    [2],
-    [1],
-    [2],
-  ]);
-  t.is(update.calls.length, 3);
-});
-
-test("broadcastMessage will not trigger unhandledMessage if at least one is a non-passive subscriber", t => {
-  const s = new Storage();
-  const emit = t.context.spy(s, "emit");
-  const init = t.context.stub(() => updateData(1));
-  const update = t.context.stub(() => updateData(2));
-  const subscribe = t.context.stub(() => ({ AAA: { passive: true } }));
-  const subscribe2 = t.context.stub(() => ({ AAA: true }));
-  const d = {
-    id: "foo",
-    init,
-    update,
-    subscribe,
-  };
-  const d2 = {
-    id: "bar",
-    init,
-    update,
-    subscribe: subscribe2,
-  };
-
-  s.createState(d);
-  s.createState(d2);
-
-  s.broadcastMessage({ tag: "AAA" });
-
-  t.deepEqual(args(emit), [
-    ["stateCreated", ["foo"], undefined, 1],
-    ["stateCreated", ["bar"], undefined, 1],
-    ["messageQueued", { tag: "AAA" }, ["@"]],
-    ["messageMatched", { tag: "AAA" }, ["foo"], true],
-    ["stateNewData", 2, ["foo"], { tag: "AAA" }],
-    ["messageMatched", { tag: "AAA" }, ["bar"], false],
-    ["stateNewData", 2, ["bar"], { tag: "AAA" }],
-  ]);
-  t.deepEqual(args(subscribe), [
-    [1],
-    [2],
-  ]);
-  t.deepEqual(args(subscribe2), [
-    [1],
-    [2],
-  ]);
-  t.is(update.calls.length, 2);
-});
-
-test("broadcastMessage will not trigger unhandledMessage if at least one is a non-passive subscriber, nested variant", t => {
-  const s = new Storage();
-  const emit = t.context.spy(s, "emit");
-  const init = t.context.stub(() => updateData(1));
-  const update = t.context.stub(() => updateData(2));
-  const subscribe = t.context.stub(() => ({ AAA: { passive: true } }));
-  const subscribe2 = t.context.stub(() => ({ AAA: true }));
-  const d = {
-    id: "foo",
-    init,
-    update,
-    subscribe,
-  };
-  const d2 = {
-    id: "bar",
-    init,
-    update,
-    subscribe: subscribe2,
-  };
-
-  const foo = s.createState(d);
-  foo.createState(d2);
-
-  s.broadcastMessage({ tag: "AAA" });
-
-  t.deepEqual(args(emit), [
-    ["stateCreated", ["foo"], undefined, 1],
-    ["stateCreated", ["foo", "bar"], undefined, 1],
-    ["messageQueued", { tag: "AAA" }, ["@"]],
-    ["messageMatched", { tag: "AAA" }, ["foo", "bar"], false],
-    ["stateNewData", 2, ["foo", "bar"], { tag: "AAA" }],
-    ["messageMatched", { tag: "AAA" }, ["foo"], true],
-    ["stateNewData", 2, ["foo"], { tag: "AAA" }],
-  ]);
-  t.deepEqual(args(subscribe), [
-    [1],
-    [2],
-  ]);
-  t.deepEqual(args(subscribe2), [
-    [1],
-    [2],
-  ]);
-  t.is(update.calls.length, 2);
-});
-
-test("broadcastMessage propagates messages in order", t => {
-  const s = new Storage();
-  const emit = t.context.spy(s, "emit");
-  const init = t.context.stub(() => updateData(1));
-  const update = t.context.stub(() => updateAndSend(2, { tag: "BBB" }));
-  const subscribe = t.context.stub(() => ({ AAA: true, BBB: true }));
-  const d = {
-    id: "foo",
-    init,
-    update,
-    subscribe,
-  };
-
-  const foo = s.createState(d);
-  foo.createState(d);
-  foo.createState(d, undefined, "bar");
-
-  s.broadcastMessage({ tag: "AAA" });
-
-  t.deepEqual(args(emit), [
-    ["stateCreated", ["foo"], undefined, 1],
-    ["stateCreated", ["foo", "foo"], undefined, 1],
-    ["stateCreated", ["foo", "bar"], undefined, 1],
-    ["messageQueued", { tag: "AAA" }, ["@"]],
-    ["messageMatched", { tag: "AAA" }, ["foo", "foo"], false],
-    ["stateNewData", 2, ["foo", "foo"], { tag: "AAA" }],
-    ["messageQueued", { tag: "BBB" }, ["foo", "foo"]],
-    ["messageMatched", { tag: "AAA" }, ["foo", "bar"], false],
-    ["stateNewData", 2, ["foo", "bar"], { tag: "AAA" }],
-    ["messageQueued", { tag: "BBB" }, ["foo", "bar"]],
-    ["messageMatched", { tag: "AAA" }, ["foo"], false],
-    ["stateNewData", 2, ["foo"], { tag: "AAA" }],
-    ["messageQueued", { tag: "BBB" }, ["foo"]],
-    ["messageMatched", { tag: "BBB" }, ["foo"], false],
-    ["stateNewData", 2, ["foo"], { tag: "BBB" }],
-    ["messageQueued", { tag: "BBB" }, ["foo"]],
-    ["messageMatched", { tag: "BBB" }, ["foo"], false],
-    ["stateNewData", 2, ["foo"], { tag: "BBB" }],
-    ["messageQueued", { tag: "BBB" }, ["foo"]],
-    ["unhandledMessage", { tag: "BBB" }, ["foo"]],
-    ["unhandledMessage", { tag: "BBB" }, ["foo"]],
-    ["unhandledMessage", { tag: "BBB" }, ["foo"]],
-  ]);
 });
